Organize debug tools
