+++ tools/debug/__init__.py
@@ -1,0 +1,12 @@
+"""Debug helper package."""
+
+from .assets import check_navbar_assets, debug_dash_asset_serving, debug_navbar_icons
+from .callbacks import debug_callback_conflicts, validate_callback_system
+
+__all__ = [
+    "debug_navbar_icons",
+    "check_navbar_assets",
+    "debug_dash_asset_serving",
+    "debug_callback_conflicts",
+    "validate_callback_system",
+]
